Store imported files by importer instance to prevent problems with once importing
If node-sass is started in watch mode or if webpack is starting node-sass
only one instance of node-sass-magic-importer is initialized and reused
for subsequential importer runs. Because of this, the store was already filled
with urls of previous runs and files would not be imported in subsequent runs.
Now for every importer run a new ID is created and the store uses the ID for
determining if a file was already imported in the current run of the importer.

Fixes #87 and #107