Prevent glob importing from failing with empty dir
In some circumstances it can be useful to already import the contents of
a directory (e.g. `@import 'objects/**/*.scss'`) although the directory
is empty, because there might be files added later in the process. Up
until now this was not possible and node-sass was throwing an error.
With this patch, empty directories are ignored and no error is thrown.

Fixes #161